separate to multiple layers
