+++ backend/get-changelog-function/src/main/resources/log4j.propeties
@@ -1,0 +1,7 @@
+log = .
+log4j.rootLogger = DEBUG, LAMBDA
+
+#Define the LAMBDA appender
+log4j.appender.LAMBDA=com.amazonaws.services.lambda.runtime.log4j.LambdaAppender
+log4j.appender.LAMBDA.layout=org.apache.log4j.PatternLayout
+log4j.appender.LAMBDA.layout.conversionPattern=%d{yyyy-MM-dd HH:mm:ss} <%X{AWSRequestId}> %-5p %c{1}:%L - %m%n
