+++ backend/pom.xml
@@ -4,7 +4,16 @@
     <modelVersion>4.0.0</modelVersion>
     <groupId>com.aws</groupId>
     <artifactId>dynamodb</artifactId>
+    <packaging>pom</packaging>
     <version>0.0.1-SNAPSHOT</version>
+    <modules>
+        <module>all-configurations-function</module>
+        <module>service-layer</module>
+        <module>create-configuration-function</module>
+        <module>delete-configuration-function</module>
+        <module>update-configuration-function</module>
+        <module>get-changelog-function</module>
+    </modules>
     <name>dynamodb</name>
     <description>Demo project for DynamoDB</description>
 
@@ -28,18 +37,7 @@
             <version>1.0.1</version>
         </dependency>
 
-        <dependency>
-            <groupId>com.amazonaws</groupId>
-            <artifactId>aws-java-sdk-dynamodb</artifactId>
-            <version>1.11.901</version>
-        </dependency>
-
         <!-- https://mvnrepository.com/artifact/com.amazonaws/aws-lambda-java-events -->
-        <dependency>
-            <groupId>com.amazonaws</groupId>
-            <artifactId>aws-lambda-java-events</artifactId>
-            <version>3.6.0</version>
-        </dependency>
     </dependencies>
 
     <build>
