+++ backend/service-layer/src/main/java/com/aws/dynamodb/manager/DynamoDBManager.java
