+++ backend/update-configuration-function/src/main/java/com/aws/dynamodb/function/DynamoUpdateConfigurationsFunction.java
@@ -1,4 +1,4 @@
-package com.aws.dynamodb.functions;
+package com.aws.dynamodb.function;
 
 import com.amazonaws.services.lambda.runtime.Context;
 import com.amazonaws.services.lambda.runtime.RequestHandler;
@@ -8,7 +8,7 @@ import java.util.Map;
 import lombok.extern.log4j.Log4j;
 
 @Log4j
-public class DynamoCreateConfigurationsFunction
+public class DynamoUpdateConfigurationsFunction
     implements RequestHandler<Map<String, Object>, Map<String, Object>> {
 
   private final DaoService daoService = DaoServiceImpl.service();
@@ -17,6 +17,6 @@ public class DynamoCreateConfigurationsFunction
   public Map<String, Object> handleRequest(Map<String, Object> robotConfiguration,
                                            Context context) {
     log.debug("Request: " + robotConfiguration);
-    return daoService.createRobotConfiguration(robotConfiguration);
+    return daoService.updateRobotConfiguration(robotConfiguration);
   }
 }
